chore: refactor and changed folder layouts
