+++ src/app/page.tsx
@@ -1,11 +1,9 @@
 import Image from "next/image";
-import { HeroPage } from "./pages/hero-page";
-import Navbar from "@/app/components/navbar";
-import ContactForm from "@/app/components/ContactUs";
-import {AboutUs} from "@/app/components/aboutUs";
-import FootNote from "@/app/components/footer";
-
-
+import { HeroPage } from "@/components/hero-page";
+import Navbar from "@/components/navbar";
+import ContactForm from "@/components/ContactUs";
+import {AboutUs} from "@/components/aboutUs";
+import FootNote from "@/components/footer";
 
 export default function Home() {
   return (
